Moved GraphViz code to it's own file
